requirements.txt
torch
transformers
pytesseract
pillow
matplotlib
numpy

